// src/app/Models/Employee.ts
export interface Employee {
  id?: number; 
  firstName: string;
  middleName: string;
  lastName: string;
  maritalStatus: string;
  dob: string; 
  mobile: string;
  address: string;
  city: string;
  state: string;
  zip: number;
  email: string;
  username: string;
  password: string;
  confirmPassword?:string
}